organizes the code and resources in the nlp module into sub-directories
