+++ src/nlp/resources/wordNetNegationPrefixes.ts
