+++ src/nlp/resources/wordNetRelations.ts
